refactor(viewer): split the Groups-tab entity tables out of treeDataBuilder
Adding IfcElectricalCircuit and IfcCondition pushed treeDataBuilder.ts to
1474 lines, past its 1466 ratchet budget. Move GROUP_ENTITY_TYPES,
GroupSubFilter, SYSTEM_GROUP_TYPES and groupMatchesSubFilter into a
sibling groupEntityTypes.ts — the module the new test file is already
named after — and re-export them from treeDataBuilder so every existing
import keeps resolving. Pure move: no table row and no branch changed.

treeDataBuilder.ts now measures 1419, so its allowlist row is LOWERED
1466 -> 1419 and ALLOWLIST_DIGEST re-pinned in the same commit. No row
was raised; the new module is 65 lines and needs no entry.
